requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
# For MySQL: uncomment below
# pymysql==1.1.0

# AI / LLM
anthropic==0.34.0
openai==1.51.0

# RAG - Embeddings & Vector Search
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4

# Document Processing
pdfplumber==0.11.4
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
